+++ src/command.ts
@@ -1,5 +1,6 @@
 export const enum Direnv {
 	reload = 'direnv.reload',
+	reset = 'direnv.reset',
 	allow = 'direnv.allow',
 	block = 'direnv.block',
 	create = 'direnv.create',